Refactor renderer to display units in cell.
